app/settings: merge `settings.js` into `meta.js`
